requirement.txt
authy==2.2.3
autopep8==1.4.4
certifi==2019.9.11
chardet==3.0.4
Django==2.2.5
django-cors-headers==3.1.1
django-silk==3.0.4
djangorestframework==3.10.3
djangorestframework-jwt==1.11.0
get==2019.4.13
gprof2dot==2017.9.19
idna==2.8
Jinja2==2.10.3
MarkupSafe==1.1.1
Pillow==6.1.0
post==2019.4.13
psycopg2==2.8.3
public==2019.4.13
pycodestyle==2.5.0
pyfcm==1.4.7
Pygments==2.4.2
PyJWT==1.7.1
PySocks==1.7.1
python-dateutil==2.8.0
pytz==2019.2
query-string==2019.4.13
request==2019.4.13
requests==2.22.0
simplejson==3.16.0
six==1.12.0
sqlparse==0.3.0
stripe==2.41.0
twilio==6.31.0
urllib3==1.25.5
celery>=3.1.18
redis>=2.10.3
django-invitations>=1.9.3
django-allauth>=0.41.0
flower>=0.9.3